seperate the topic from orignal Queue service
