+++ src/subscriber/Services/Queues/QueueClientFactory.cs
@@ -34,21 +34,3 @@ public class QueueClientFactory(IEnumerable<IQueueClient> clients) : IQueueClien
         _ => throw new ArgumentException($"Unknown client type: {client.GetType().Name}")
     };
 }
-
-// Topic interfaces and factory
-public interface ITopicClientFactory
-{
-    ITopicClient GetTopicClient();
-}
-
-public class TopicClientFactory : ITopicClientFactory
-{
-    private readonly ITopicClient _topicClient;
-
-    public TopicClientFactory(ITopicClient topicClient)
-    {
-        _topicClient = topicClient ?? throw new ArgumentNullException(nameof(topicClient));
-    }
-
-    public ITopicClient GetTopicClient() => _topicClient;
-}
