+++ src/subscriber/Services/Topics/Azure/AzureServiceBusTopic.cs
@@ -1,99 +1,14 @@
 using System.Text;
 using Azure.Messaging.ServiceBus;
 using Azure.Messaging.ServiceBus.Administration;
-using Microsoft.Extensions.Logging; // Assuming standard Microsoft logging
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using subscriber.Services.Queues.Exceptions;
 using subscriber.Configuration.ServiceBus;
-using Microsoft.Extensions.Options;
+using subscriber.Services.Topics;
 
-namespace subscriber.Services.Queues.Azure
+namespace subscriber.Services.Topics.Azure
 {
-	/// <summary>
-	/// Represents a message received from a Service Bus topic subscription
-	/// Implements IQueueMessage and holds the original message for settlement.
-	/// </summary>
-	public class TopicSubscriptionMessage(ServiceBusReceivedMessage message, string subscriptionName) : IQueueMessage
-	{
-		// Keep the original message private but accessible internally if needed,
-		// or provide a specific method/property for the client implementation
-		// For simplicity here, we make it accessible via a property for the client.
-		internal ServiceBusReceivedMessage OriginalMessage { get; } = message ?? throw new ArgumentNullException(nameof(message));
-
-		public string MessageId => OriginalMessage.MessageId;
-
-		// Provide Body as UTF8 string for convenience
-		public string Body => Encoding.UTF8.GetString(OriginalMessage.Body.ToArray());
-
-		// Expose raw bytes
-		public ReadOnlyMemory<byte> BodyBytes => OriginalMessage.Body;
-
-		public DateTime EnqueuedTime => OriginalMessage.EnqueuedTime.UtcDateTime;
-		public string ReceiptHandle => OriginalMessage.LockToken;
-		public uint DeliveryCount => (uint)OriginalMessage.DeliveryCount;
-		public IDictionary<string, string> Properties =>
-				OriginalMessage.ApplicationProperties?.ToDictionary(
-						p => p.Key,
-						p => p.Value?.ToString() ?? string.Empty)
-				?? new Dictionary<string, string>();
-
-		// Store the subscription name the message came from
-		public string SubscriptionName { get; } = subscriptionName ?? throw new ArgumentNullException(nameof(subscriptionName));
-	}
-
-	/// <summary>
-	/// Manages Azure Service Bus topics and subscriptions
-	/// </summary>
-	public interface ITopicClient
-	{
-		/// <summary>
-		/// Initialize the topic client (create topic/subscriptions if needed)
-		/// </summary>
-		Task InitializeAsync(CancellationToken cancellationToken);
-
-		/// <summary>
-		/// Publish a message to the topic
-		/// </summary>
-		Task PublishMessageAsync(string messageBody, IDictionary<string, string>? properties = null, CancellationToken cancellationToken = default);
-
-		/// <summary>
-		/// Receive messages from a specific subscription.
-		/// Messages received via this method must be explicitly completed, abandoned, or dead-lettered
-		/// using the corresponding methods below before their lock expires.
-		/// </summary>
-		/// <param name="subscriptionName">The name of the subscription to receive from.</param>
-		/// <param name="maxMessages">The maximum number of messages to receive.</param>
-		/// <param name="maxWaitTime">The maximum time to wait for messages to arrive. Null means wait indefinitely (or until cancellation).</param>
-		/// <param name="cancellationToken">Cancellation token.</param>
-		/// <returns>A collection of messages received.</returns>
-		Task<IEnumerable<TopicSubscriptionMessage>> ReceiveMessagesFromSubscriptionAsync(
-				string subscriptionName,
-				int maxMessages,
-				TimeSpan? maxWaitTime, // Renamed from visibilityTimeout, made nullable
-				CancellationToken cancellationToken);
-
-		/// <summary>
-		/// Complete a message (mark as processed).
-		/// </summary>
-		/// <param name="message">The message received from ReceiveMessagesFromSubscriptionAsync.</param>
-		/// <param name="cancellationToken">Cancellation token.</param>
-		Task CompleteMessageAsync(TopicSubscriptionMessage message, CancellationToken cancellationToken);
-
-		/// <summary>
-		/// Abandon a message (return to subscription for later processing).
-		/// </summary>
-		/// <param name="message">The message received from ReceiveMessagesFromSubscriptionAsync.</param>
-		/// <param name="cancellationToken">Cancellation token.</param>
-		Task AbandonMessageAsync(TopicSubscriptionMessage message, CancellationToken cancellationToken);
-
-		/// <summary>
-		/// Dead-letter a message (move to Dead Letter Queue).
-		/// </summary>
-		/// <param name="message">The message received from ReceiveMessagesFromSubscriptionAsync.</param>
-		/// <param name="reason">The reason for dead-lettering.</param>
-		/// <param name="cancellationToken">Cancellation token.</param>
-		Task DeadLetterMessageAsync(TopicSubscriptionMessage message, string reason, CancellationToken cancellationToken);
-	}
-
 	/// <summary>
 	/// Implementation of Service Bus topic client
 	/// </summary>
